Add files via upload
I recently worked on a project involving gender classification using machine learning techniques. The goal of the project was to accurately predict an individual’s gender based on certain features. We used a dataset that provided data on personal preferences and trained our model to achieve an accuracy . The results were promising and showed the potential for further research in this area. #datascience #machinelearning #genderclassification